add option to have separate predtimechart config
At the moment, it's assumed that the configuration for predtimechart
always lives in the hub under `hub-config/predtimechart-config.yml`.

This change proposes an _optional argument_ that allows the user to
specify the path to this config file so the following scenarios could be
achieved:

 - a 4-letter extension `hub-config/predtimechart-config.yaml`
 - a hiddent file called `hub-config/.predtimechart-config.yml`
 - a file in a different repository so that the dashboard configuration
   can be completely separate from the hub workings

NOTE: I _do not_ know how click works and I do not know how to create an
optional argument (most python people look down upon us lowly R folk and
tut tut about our weakly-typed functions, to which I say: fair enough).

This will _likely fail_ because of my lack of knowledge of python
workings.